+++ app/src/components/list-albums/list-albums.js
@@ -10,6 +10,7 @@ controller.$inject = ['albumService'];
 
 function controller(albumService){
   this.styles = styles;
+
   albumService.getAll()
     .then(albums=>this.albums=albums)
     .catch(err=>console.log(err));
@@ -17,6 +18,17 @@ function controller(albumService){
   this.add = albumToAdd=>{
     albumService.add(albumToAdd)
       .then(addedAlbum=>this.albums.push(addedAlbum))
+      .catch(err=>console.log(err));
+  };
+
+  this.remove = albumId=>{
+    albumService.remove(albumId)
+      .then(deleted=>{
+        const index = this.albums.findIndex(album=>album._id === deleted._id);
+        if(index !== -1){
+          this.albums.splice(index, 1);
+        }
+      })
       .catch(err=>console.log(err));
   };
 }