+++ app/src/services/album-service.js
@@ -1,14 +1,27 @@
-albumService.$inject = ['$http', 'apiUrl'];
+albumService.$inject = ['$http', 'apiUrl'/*, '$cacheFactory'*/];
 
-export default function albumService($http, apiUrl){
+//TODO: revisit caching on GETs
+
+export default function albumService($http, apiUrl/*, $cacheFactory*/){
+  //const cache = $cacheFactory.get('$http');
+
   return {
     getAll(){
-      return $http.get(`${apiUrl}/albums`)
+      return $http.get(`${apiUrl}/albums`/*, {cache: true}*/)
         .then(response=>response.data)
         .catch(err=>console.log(err));
     },
     add(album){
+      //cache.remove(`${apiUrl}/albums`);
+
       return $http.post(`${apiUrl}/albums`, album)
+        .then(response=>response.data)
+        .catch(err=>console.log(err));
+    },
+    remove(albumId){
+      //cache.remove(`${apiUrl}/albums`);
+  
+      return $http.delete(`${apiUrl}/albums/${albumId}`)
         .then(response=>response.data)
         .catch(err=>console.log(err));
     }